terraform/scripts/cloud_function/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-dataproc==5.*
